requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pydantic==2.3.0 